Add stress test for paper exchange and update auto-release workflow to ignore it
